Tests/LibWeb/Text/input/abspos-relayout-fixed-descendant-fallback.html

<!DOCTYPE html>
<script src="include.js"></script>
<style>
    body { margin: 0; }
    #parent { position: relative; width: 360px; height: 220px; }
    #boundary {
        position: absolute;
        left: 100px;
        top: 100px;
        width: 200px;
        height: 120px;
        overflow: hidden;
    }
    #fixed {
        position: fixed;
        left: 30px;
        top: 20px;
        width: 40px;
        height: 25px;
    }
    #sibling { width: 80px; height: 20px; }
</style>
<div id="parent">
    <div id="boundary">
        <div id="fixed"></div>
        <div id="sibling"></div>
    </div>
</div>
<script>
    function rectToString(rect) {
        return `${rect.left},${rect.top},${rect.width},${rect.height}`;
    }

    asyncTest(async (done) => {
        document.body.offsetHeight;

        const fixedBefore = document.getElementById("fixed").getBoundingClientRect();
        const partialBefore = internals.partialLayoutCount();
        const fullBefore = internals.fullLayoutCount();
        document.getElementById("sibling").style.height = "70px";
        document.body.offsetHeight;
        const partialDelta = internals.partialLayoutCount() - partialBefore;
        const fullDelta = internals.fullLayoutCount() - fullBefore;

        const fixedAfter = document.getElementById("fixed").getBoundingClientRect();
        const sibling = document.getElementById("sibling").getBoundingClientRect();
        println(`fixed-before=${rectToString(fixedBefore)}`);
        println(`fixed-after=${rectToString(fixedAfter)}`);
        println(`sibling=${rectToString(sibling)}`);
        println(`partial-layouts=${partialDelta} full-layouts=${fullDelta}`);
        println(rectToString(fixedBefore) === "30,20,40,25"
            && rectToString(fixedBefore) === rectToString(fixedAfter)
            && rectToString(sibling) === "100,100,80,70"
            && partialDelta === 0 && fullDelta === 1 ? "PASS" : "FAIL");
        done();
    });
</script>
